fix bug while loading 3D model, it was the type of the indices offset

+++ sandbox_useful/inc/sandbox_useful/buffer/vertex.hpp
@@ -76,7 +76,7 @@ namespace buffer{
 
         uint32_t _indices_count;
         vertex_description _description;
-        uint8_t _indices_offset;
+        uint32_t _indices_offset;
     };
 
 #include "sandbox_useful/buffer/vertex.tpp"
